backend/requirements.txt
fastapi==0.135.1
python-multipart==0.0.22
uvicorn[standard]==0.41.0
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pandas==3.0.1
requests==2.32.5
bcrypt==5.0.0
PyJWT==2.12.1
